creditcoin: move helpers into separate module
