+++ pallets/creditcoin/src/helpers.rs
@@ -1,0 +1,46 @@
+use crate::{
+	pallet::*,
+	types::{Address, AddressId},
+};
+use frame_system::pallet_prelude::*;
+use sp_runtime::RuntimeAppPublic;
+use sp_std::prelude::*;
+
+#[allow(unused_macros)]
+macro_rules! try_get {
+	($storage: ident <$t: ident>, $key: expr, $err: ident) => {
+		crate::pallet::$storage::<$t>::try_get($key).map_err(|()| crate::pallet::Error::<$t>::$err)
+	};
+}
+
+macro_rules! try_get_id {
+	($storage: ident <$t: ident>, $key: expr, $err: ident) => {
+		<crate::pallet::$storage<$t> as DoubleMapExt<_, _, _, _, _, _, _, _, _, _>>::try_get_id(
+			$key,
+		)
+		.map_err(|()| crate::pallet::Error::<$t>::$err)
+	};
+}
+
+impl<T: Config> Pallet<T> {
+	pub fn block_number() -> BlockNumberFor<T> {
+		<frame_system::Pallet<T>>::block_number()
+	}
+	pub fn get_address(address_id: &AddressId<T::Hash>) -> Result<Address<T::AccountId>, Error<T>> {
+		Self::addresses(&address_id).ok_or(Error::<T>::NonExistentAddress)
+	}
+
+	pub fn authority_id() -> Option<T::AccountId> {
+		let local_keys = crate::crypto::Public::all()
+			.into_iter()
+			.map(|p| sp_core::sr25519::Public::from(p).into())
+			.collect::<Vec<T::FromAccountId>>();
+
+		log::trace!("{:?}", local_keys);
+
+		Authorities::<T>::iter_keys().find_map(|auth| {
+			let acct = auth.clone().into();
+			local_keys.contains(&acct).then(|| auth)
+		})
+	}
+}
