+++ pallets/creditcoin/src/lib.rs
@@ -2,9 +2,8 @@
 
 extern crate alloc;
 
-use frame_system::pallet_prelude::*;
 pub use pallet::*;
-use sp_runtime::{KeyTypeId, RuntimeAppPublic};
+use sp_runtime::KeyTypeId;
 use sp_std::prelude::*;
 
 #[cfg(test)]
@@ -16,6 +15,8 @@ mod tests;
 #[cfg(feature = "runtime-benchmarks")]
 mod benchmarking;
 
+#[macro_use]
+mod helpers;
 mod ocw;
 mod types;
 
@@ -44,22 +45,6 @@ pub mod crypto {
 }
 
 pub type BalanceFor<T> = <T as pallet_balances::Config>::Balance;
-
-#[allow(unused_macros)]
-macro_rules! try_get {
-	($storage: ident <$t: ident>, $key: expr, $err: ident) => {
-		crate::pallet::$storage::<$t>::try_get($key).map_err(|()| crate::pallet::Error::<$t>::$err)
-	};
-}
-
-macro_rules! try_get_id {
-	($storage: ident <$t: ident>, $key: expr, $err: ident) => {
-		<crate::pallet::$storage<$t> as DoubleMapExt<_, _, _, _, _, _, _, _, _, _>>::try_get_id(
-			$key,
-		)
-		.map_err(|()| crate::pallet::Error::<$t>::$err)
-	};
-}
 
 #[frame_support::pallet]
 pub mod pallet {
@@ -717,28 +702,5 @@ pub mod pallet {
 
 			Ok(())
 		}
-	}
-}
-
-impl<T: Config> Pallet<T> {
-	pub fn block_number() -> BlockNumberFor<T> {
-		<frame_system::Pallet<T>>::block_number()
-	}
-	pub fn get_address(address_id: &AddressId<T::Hash>) -> Result<Address<T::AccountId>, Error<T>> {
-		Self::addresses(&address_id).ok_or(Error::<T>::NonExistentAddress)
-	}
-
-	fn authority_id() -> Option<T::AccountId> {
-		let local_keys = crypto::Public::all()
-			.into_iter()
-			.map(|p| sp_core::sr25519::Public::from(p).into())
-			.collect::<Vec<T::FromAccountId>>();
-
-		log::trace!("{:?}", local_keys);
-
-		Authorities::<T>::iter_keys().find_map(|auth| {
-			let acct = auth.clone().into();
-			local_keys.contains(&acct).then(|| auth)
-		})
 	}
 }
